# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ExpensesManagementApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Person": {
      "post": {
        "tags": [
          "Person"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PersonPostRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Person"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Person"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PersonDeleteRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Person"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PersonPutRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Person/TotalBalance": {
      "get": {
        "tags": [
          "Person"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Transaction": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TransactionPostRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Transaction"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "PersonDeleteRequest": {
      "type": "object",
      "properties": {
        "idsToDelete": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        }
      }
    },
    "PersonPostRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "age": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "PersonPutRequest": {
      "required": [
        "id"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "age": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "TransactionPostRequest": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "value": {
          "format": "double",
          "type": "number"
        },
        "type": {
          "$ref": "#/definitions/TransactionType"
        },
        "personId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "TransactionType": {
      "format": "int32",
      "enum": [
        0,
        1
      ],
      "type": "integer"
    }
  }
}